jobs/requirements.txt
requests==2.32.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0

requests==2.32.3
